Add file rotation
Also:
- Name files with unix epoch seconds timestamps only
- Rename RUSTCASK_MAX_FILE_SIZE to RUSTCASK_ROTATE_ACTIVE_FILE_AFTER_BYTES to better indicate what it does
- Add root path and Map of readers to rotated files to RustcaskDatabase struct
- Return Vec of deleted keys from rebuild_keydir to correctly build database on startup from multiple files

+++ src/settings.rs
@@ -3,11 +3,11 @@ use std::path::PathBuf;
 #[derive(Debug)]
 pub struct Settings {
     pub file_path: PathBuf,
-    pub max_file_size: u8, // Max file size in MB, defaults to 4 MB
+    pub max_file_size: u64,
 }
 
 impl Settings {
-    pub fn new(file_path: PathBuf, max_file_size: u8) -> Self {
+    pub fn new(file_path: PathBuf, max_file_size: u64) -> Self {
         Self {
             file_path,
             max_file_size,
